% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{c3d_example}
\alias{c3d_example}
\title{Get path to c3dr example}
\usage{
c3d_example()
}
\value{
A character vector with the absolute file path of the example
file.
}
\description{
Return the file path for an example data files within the \code{c3dr} package.
}
\details{
The test data file contains a short recording of human walking using a
full-body model. The test data includes analog channels (e.g., EMG) and data
from two force platforms. The recording was made with a Qualisys motion
capture system.

The file is taken from \url{https://github.com/pyomeca/ezc3d-testFiles} under a
GPL-3.0 license.
}
\examples{
c3d_example()

}
