% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{c3d_longer}
\alias{c3d_longer}
\title{Convert wide to long data for c3d points}
\usage{
c3d_longer(x)
}
\arguments{
\item{x}{A data.frame of class \code{c3d} with c3d point data with three
columns (x, y, z) per point (wide format).}
}
\value{
A data.frame with one column per point (long format). It has n*3 rows
and k columns, with n as the number of recorded frames and k as the
number of recorded points.
}
\description{
Convert from wide representation of data (3 columns per point) to long data
(1 column per point) in c3d point data.
}
\examples{
# Import example data
d <- c3d_read(c3d_example())
# get point data in wide format
w <- c3d_data(d, format = "wide")
# convert to long data
l <- c3d_longer(w)
head(l)

}
