% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{format.c3d}
\alias{format.c3d}
\title{Formatting c3d objects}
\usage{
\method{format}{c3d}(x, ...)
}
\arguments{
\item{x}{A \code{list} of the class \code{c3d} to be formatted.}

\item{...}{empty argument, currently not used.}
}
\value{
A character string with basic information for the c3d object.
}
\description{
Formatting method for c3d objects
}
\examples{
# Import example data
d <- c3d_read(c3d_example())

format(d)
}
