% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.c3d}
\alias{print.c3d}
\title{Printing c3d objects}
\usage{
\method{print}{c3d}(x, ...)
}
\arguments{
\item{x}{A \code{list} of the class \code{c3d} to be printed.}

\item{...}{empty argument, currently not used.}
}
\value{
The function prints basic information for the c3d object and returns
it invisibly.
}
\description{
Printing method for c3d objects
}
\details{
Prints c3d objects by calling \code{\link[=format.c3d]{format.c3d()}}.
}
\examples{
# Import example data
d <- c3d_read(c3d_example())

print(d)
}
