% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{c3d_analog}
\alias{c3d_analog}
\title{Get analog data from a c3d file}
\usage{
c3d_analog(x)
}
\arguments{
\item{x}{A c3d object, as imported by \code{c3d_read()}.}
}
\value{
A data.frame with with n rows and m columns, where n is the number of
frames recorded times the number of analog subframes per frame, and m as
the number of recorded analog channels.
}
\description{
Get the analog data in a data.frame
}
