% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{c3d_data}
\alias{c3d_data}
\title{Get point data from a c3d file}
\usage{
c3d_data(x)
}
\arguments{
\item{x}{A c3d object, as imported by \code{c3d_read()}.}
}
\value{
A data.frame with three columns (x, y, z) per point.
}
\description{
Get the point data in a data.frame
}
