% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{c3d_data}
\alias{c3d_data}
\title{Get point data from a c3d object}
\usage{
c3d_data(x, format = "wide")
}
\arguments{
\item{x}{A c3d object, as imported by \code{\link{c3d_read}}.}

\item{format}{Either "wide" (default), "long", or "longest" to determine the
format of the resulting data frame. The wide format has Long format has
three columns per point (x,y,z), the long format has three rows per frame
per point (x,y,z). The longest format has one row per data entry.}
}
\value{
A data frame of class \code{c3d_data} with the c3d point data. The
structure of the data frame depends on the 'format' argument.
}
\description{
Get the point data of an c3d object in a data frame.
}
\details{
The point data of imported c3d objects in \code{c3dr} is saved as a list of
lists. This is good for internal handling, but for analysis a table format (a
data frame) is often more convenient.\code{c3d_data()} returns the point data
from an imported c3d object as a data frame.

Analyses of data frames may require them to have different formats. For
\code{c3d_data} output, different data formats (\code{"wide"}, \code{"long"},
\code{"longest"}) are available. See the \code{format} argument for more
details. You can convert between different formats with
\code{\link{c3d_convert}}.
}
\examples{
# Import example data
d <- c3d_read(c3d_example())

# wide format (default)
w <- c3d_data(d)
head(w)

# long format
l <- c3d_data(d, format = "long")
head(l)

# longest format
ll <- c3d_data(d, format = "longest")
head(ll)
}
