% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{c3d_example}
\alias{c3d_example}
\title{Get path to c3dr example}
\usage{
c3d_example()
}
\value{
A character vector with the absolute file path of the example
file.
}
\description{
\code{c3d_example} returns the file path for an example data files within the
\code{c3dr} package.
}
\examples{
c3d_example()

}
