% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{c3d_longer}
\alias{c3d_longer}
\title{Convert wide to long data for c3d points}
\usage{
c3d_longer(x)
}
\arguments{
\item{x}{A data.frame with c3d point data with three columns (x, y, z) per
point.}
}
\value{
A data.frame with one column per point. It has n*3 rows
and k columns, with n as the number of recorded frames and k as the
number of recorded points.
}
\description{
Convert from wide representation of data (3 columns per point) to long data
(1 column per point) in c3d point data
}
