% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{c3d_read}
\alias{c3d_read}
\title{Read a c3d file in R}
\usage{
c3d_read(file)
}
\arguments{
\item{file}{A string with the path of a c3d file.}
}
\value{
A list of class c3d.
}
\description{
\code{c3dr_read()} imports a c3d file using the C++ ezx3d library.
}
