% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdata.R
\name{c3d_setdata}
\alias{c3d_setdata}
\title{Write data to a c3d object}
\usage{
c3d_setdata(object, newdata = NULL, newanalog = NULL)
}
\arguments{
\item{object}{A \code{c3d} object to be modified.}

\item{newdata}{The new point data that should be written to the \code{c3d}
object. Usually a data frame of the class \code{c3d_data} as it is
generated by \code{\link{c3d_data}}.}

\item{newanalog}{The new analog data that should be written to the \code{c3d}
object. Usually a data frame of the class \code{c3d_analog} as it is
generated by \code{\link{c3d_analog}}.}
}
\value{
The modified c3d object.
}
\description{
Set new data to an existing c3d object
}
\examples{
# Import example data
d <- c3d_read(c3d_example())

# remove last frame from point data and analog data (10 subframes for analog)
d_cut <- c3d_data(d)[-340,]
a_cut <- c3d_analog(d)[-c(3391:3400),]

# write the new c3d object
d_new <- c3d_setdata(d, newdata = d_cut, newanalog = a_cut)
d_new

}
