% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setdata.R
\name{c3d_setdata}
\alias{c3d_setdata}
\title{Write data to a c3d object}
\usage{
c3d_setdata(object, newdata = NULL, newanalog = NULL)
}
\arguments{
\item{object}{A \code{c3d} object to be modified.}

\item{newdata}{The new point data that should be written to the \code{c3d}
object.}

\item{newanalog}{The new analog data that should be written to the \code{c3d}
object.}
}
\value{
The modified c3d object
}
\description{
Set new data to an existing c3d object
}
